replace matching with the refactored matching package
