+++ matching/pricelevel_test.go
@@ -10,21 +10,21 @@ import (
 )
 
 func TestGetPriceLevel(t *testing.T) {
-	side := &OrderBookSide{}
+	side := &SellSide{}
 	assert.Equal(t, 0, len(side.levels))
-	side.getPriceLevel(100, types.Side_Sell)
+	side.getPriceLevel(100)
 	assert.Equal(t, 1, len(side.levels))
 
-	side.getPriceLevel(110, types.Side_Sell)
+	side.getPriceLevel(110)
 	assert.Equal(t, 2, len(side.levels))
 
-	side.getPriceLevel(100, types.Side_Sell)
+	side.getPriceLevel(100)
 	assert.Equal(t, 2, len(side.levels))
 }
 
 func TestAddAndRemoveOrdersToPriceLevel(t *testing.T) {
-	side := &OrderBookSide{}
-	l := side.getPriceLevel(100, types.Side_Sell)
+	side := &SellSide{}
+	l := side.getPriceLevel(100)
 	order := &types.Order{
 		MarketID:    "testOrderBook",
 		PartyID:     "A",
@@ -54,8 +54,8 @@ func TestUncross(t *testing.T) {
 	logger := logging.NewTestLogger()
 	defer logger.Sync()
 
-	side := &OrderBookSide{}
-	l := side.getPriceLevel(100, types.Side_Sell)
+	side := &SellSide{}
+	l := side.getPriceLevel(100)
 	passiveOrder := &types.Order{
 		MarketID:    "testOrderBook",
 		PartyID:     "A",
