+++ matching/side_test.go
@@ -9,11 +9,12 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func getTestSide() *OrderBookSide {
-	return &OrderBookSide{
-		log:    logging.NewTestLogger(),
-		levels: []*PriceLevel{},
-	}
+func getTestSide() *SellSide {
+	return NewSellSide(logging.NewTestLogger())
+}
+
+func getTestBuySide() *BuySide {
+	return NewBuySide(logging.NewTestLogger())
 }
 
 func TestMemoryAllocationPriceLevelRemoveOrder(t *testing.T) {
@@ -29,7 +30,7 @@ func TestMemoryAllocationPriceLevelRemoveOrder(t *testing.T) {
 		TimeInForce: types.Order_GTC,
 	}
 	// add the order to the side
-	side.addOrder(o, types.Side_Sell)
+	side.AddOrder(o)
 	assert.Len(t, side.levels, 1)
 
 	o2 := &types.Order{
@@ -44,7 +45,7 @@ func TestMemoryAllocationPriceLevelRemoveOrder(t *testing.T) {
 	}
 
 	// add the order to the side
-	side.addOrder(o2, types.Side_Sell)
+	side.AddOrder(o2)
 	assert.Len(t, side.levels, 2)
 
 	// remove it and check the length of the array
@@ -57,61 +58,61 @@ func TestMemoryAllocationGetPriceLevelReturnAPriceLevelIfItAlreadyExists(t *test
 	// test for a sell side
 	side := getTestSide()
 	assert.Len(t, side.levels, 0)
-	pl := side.getPriceLevel(100, types.Side_Sell)
+	pl := side.getPriceLevel(100)
 	assert.Len(t, side.levels, 1)
 	assert.NotNil(t, pl)
-	pl = side.getPriceLevel(101, types.Side_Sell)
+	pl = side.getPriceLevel(101)
 	assert.Len(t, side.levels, 2)
 	assert.NotNil(t, pl)
-	pl = side.getPriceLevel(102, types.Side_Sell)
+	pl = side.getPriceLevel(102)
 	assert.Len(t, side.levels, 3)
 	assert.NotNil(t, pl)
-	pl = side.getPriceLevel(103, types.Side_Sell)
+	pl = side.getPriceLevel(103)
 	assert.Len(t, side.levels, 4)
 	assert.NotNil(t, pl)
-	pl = side.getPriceLevel(104, types.Side_Sell)
+	pl = side.getPriceLevel(104)
 	assert.Len(t, side.levels, 5)
 	assert.NotNil(t, pl)
 
 	// get existing one in bounds now
-	pl = side.getPriceLevel(102, types.Side_Sell)
+	pl = side.getPriceLevel(102)
 	assert.Len(t, side.levels, 5)
 	assert.NotNil(t, pl)
-	pl = side.getPriceLevel(100, types.Side_Sell)
+	pl = side.getPriceLevel(100)
 	assert.Len(t, side.levels, 5)
 	assert.NotNil(t, pl)
-	pl = side.getPriceLevel(104, types.Side_Sell)
+	pl = side.getPriceLevel(104)
 	assert.Len(t, side.levels, 5)
 	assert.NotNil(t, pl)
 
 	// test for a buy side
-	side = getTestSide()
-	assert.Len(t, side.levels, 0)
-	pl = side.getPriceLevel(100, types.Side_Buy)
-	assert.Len(t, side.levels, 1)
+	bside := getTestBuySide()
+	assert.Len(t, bside.levels, 0)
+	pl = bside.getPriceLevel(100)
+	assert.Len(t, bside.levels, 1)
 	assert.NotNil(t, pl)
-	pl = side.getPriceLevel(101, types.Side_Buy)
-	assert.Len(t, side.levels, 2)
+	pl = bside.getPriceLevel(101)
+	assert.Len(t, bside.levels, 2)
 	assert.NotNil(t, pl)
-	pl = side.getPriceLevel(102, types.Side_Buy)
-	assert.Len(t, side.levels, 3)
+	pl = bside.getPriceLevel(102)
+	assert.Len(t, bside.levels, 3)
 	assert.NotNil(t, pl)
-	pl = side.getPriceLevel(103, types.Side_Buy)
-	assert.Len(t, side.levels, 4)
+	pl = bside.getPriceLevel(103)
+	assert.Len(t, bside.levels, 4)
 	assert.NotNil(t, pl)
-	pl = side.getPriceLevel(104, types.Side_Buy)
-	assert.Len(t, side.levels, 5)
+	pl = bside.getPriceLevel(104)
+	assert.Len(t, bside.levels, 5)
 	assert.NotNil(t, pl)
 
 	// get existing one in bounds now
-	pl = side.getPriceLevel(102, types.Side_Buy)
-	assert.Len(t, side.levels, 5)
+	pl = bside.getPriceLevel(102)
+	assert.Len(t, bside.levels, 5)
 	assert.NotNil(t, pl)
-	pl = side.getPriceLevel(100, types.Side_Buy)
-	assert.Len(t, side.levels, 5)
+	pl = bside.getPriceLevel(100)
+	assert.Len(t, bside.levels, 5)
 	assert.NotNil(t, pl)
-	pl = side.getPriceLevel(104, types.Side_Buy)
-	assert.Len(t, side.levels, 5)
+	pl = bside.getPriceLevel(104)
+	assert.Len(t, bside.levels, 5)
 	assert.NotNil(t, pl)
 }
 
@@ -128,7 +129,7 @@ func TestMemoryAllocationPriceLevelUncrossSide(t *testing.T) {
 		TimeInForce: types.Order_GTC,
 	}
 	// add the order to the side
-	side.addOrder(o, types.Side_Sell)
+	side.AddOrder(o)
 	assert.Len(t, side.levels, 1)
 
 	o2 := &types.Order{
@@ -143,7 +144,7 @@ func TestMemoryAllocationPriceLevelUncrossSide(t *testing.T) {
 	}
 
 	// add the order to the side
-	side.addOrder(o2, types.Side_Sell)
+	side.AddOrder(o2)
 	assert.Len(t, side.levels, 2)
 
 	aggressiveOrder := &types.Order{
@@ -157,6 +158,6 @@ func TestMemoryAllocationPriceLevelUncrossSide(t *testing.T) {
 		TimeInForce: types.Order_GTC,
 	}
 
-	side.uncross(aggressiveOrder)
+	side.Uncross(aggressiveOrder)
 	assert.Len(t, side.levels, 1)
 }
